If you call a ModelModel mock (eg., "sim_inst.Model()") it no longer returns the the first element of the entity list but the complete entity list.
Originally, I had the assumption, that creating one model instance will always result in one entity, but this is not the case for, e.g., a power grid where the grid consists of multiple buses and branches. However, the current implementation is still not optimal. If you create multiple instances in a list (like "[sim_inst.Model() for i in range(3)]" you would end up with a lists of lists of entities. Maybe I have to introduce some kind of hierarchical model, where the buses and branches would be children of a PowerGrid entity.