First working version of new data flow
The caches are not yet pruned.

Same-time loops are not supported by the new system yet, it might be a
good idea to take a more principled approach (like super-dense time)
here.

Async requests are not compatible with the new system. As they are
scheduled for removal anyway, it might not be worth adding (or keeping)
the separate code path.