Rework scheduling based on dense time
The scheduler has received an overhaul. It is now goverened by the
following principle: There are two numbers that approach each other:
- The simulator's currently-scheduled earliest next step. This can
  go down as earlier steps are scheduled by other simulators.
- The simulator's potentially-scheduled earliest next step (called its
  progress), based on the events anywhere is the simulation. This can
  go up when events elsewhere happen to not actually schedule events
  for this simulator.
Once both of these numbers agree, the next step is settled. The
simulator will then wait for its dependencies (and its dependants
in case of lazy stepping or async requests) to step and produce data.