Reorganize execution graph
The new execution graph has (sim_id, dense_time) pairs as nodes that
represent the actual dense times when the simulators are run. (And
reduces the amount of string parsing required while writing the graph.)
There are a couple of scenarios where the expected graph had to be
adjusted. (Previously, the second time components counted repeat
invocations of the simulator; but a first invocation occasionally occurs
at micro-step 1, which is now reflected.)