Introduce InOrOutSet for attribute handling
InOrOutSets can either represent a finite set of elements or the
complement of such a set (i.e. all but a finite number of elements).
The latter is called an OutSet.

This is useful for handling attrs in the presence of any_inputs=True.
In this case, the trigger or non-trigger inputs are represented by such
an OutSet, namely all potential attrs except those explicit marked as
the other type. By using OutSets, tests for any_inputs can be avoided
in the rest of the code. (This will be added in a future commit.)