Introduce AsyncWorld
AsyncWorld is like World, except that all methods that used the event
loop in their implementation are now async, instead. This allows them to
be used in an existing event loop. Similarly, AsyncModelFactory any
AsyncModelMock are introduced, as their non-async counterparts also
explicitly called into the event loop.

To keep compatibility with existing scenario scripts, World,
ModelFactory, and ModelMock are re-introduced as proxy objects that
schedule the coroutines of their async versions in an event loop
managed by World as before.