Improvements to documentation links
Fixes for many incorrect references in the docs. The main thing that
remains is fixing references to type aliases and typed dicts in type
annotations. There currently seems to be no good solution for this
because sphinx automatically references types in type annotations
by the role :py:class:, but type alias have the role :py:data: and
typed dicts have the role :py:typeddict: (when using autotypeddict
from sphinx_toolbox). There seems to be some discussion on this in
sphinx's GitHub issues, so hopefully there will be a solution soon.

For now, we keep the warnings as a reminder.