Fix scheduler tests
The introduction of the warning if shutdown is not performed revealed
the the scheduler tests also did not shut down the world properly, due
to the fact that pytest-asyncio "steals" the world's event loop.

The fix is to use the AsyncWorld in the tests directly. This is correct
anyway, as the scheduler works on the level of AsyncWorld.